requirements.txt
fastapi[standard]==0.115.8
uvicorn==0.34.0
APScheduler==3.11.0
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
requests==2.32.3
